feat: speed-controlled fans
This implements the discovery side for speed-controlled fans. Here we no
longer need to emulate speed changes by changing modes and there is a
readily-usable onOff cluster for state. Fans now need to be either mode-
or speed-controlled, otherwise an assert is hit. No such fans currently
exist in the Z2M codebase and they cannot be reasonably controlled
anyways.